DONE: Constructors -- need for speed.
this exercise has been the least clear of all that I've done on exercism. Very poorly worded instructions paried with nearly no direction on what topics need to be understood to complete the task. Speed is equated to distance? distance is a measure of speed over time, speed isn't synonymous with distance, and it's strange that they frame it as such since all of these exercises are just middleschool math problems. discouraging problem set.